requirements.txt
# Django и основные зависимости
Django==5.1
python-decouple==3.8
asgiref==3.8.1

# Оптимизация статики
whitenoise==6.7.0

# База данных — PostgreSQL
psycopg2-binary>=2.9.9
# PyMySQL>=1.1.0  # Устарело: перешли на PostgreSQL

# Статические файлы и медиа
Pillow==10.4.0
django-ckeditor==6.7.1

# Админ панель
django-jazzmin==2.6.0

# Деревья и теги
django-mptt==0.16.0
django-taggit==5.0.1

# Redis + Celery (фоновые задачи + кэш)
redis>=5.0.1
django-redis>=5.4.0
celery>=5.3.0
celery[redis]>=5.3.0
django-celery-beat>=2.6.0
django-celery-results>=2.5.1

# Планировщик задач (apscheduler оставляем для совместимости)
django-apscheduler==0.7.0

# Таблицы
django-tables2==2.7.5

# Telegram
telepot==12.7

# Утилиты
unidecode==1.3.7

# Оптимизация производительности
django-debug-toolbar==4.3.0
django-extensions==3.2.3

# Безопасность
django-cors-headers==4.3.1
django-csp==3.7

# Мониторинг и логирование
sentry-sdk==1.40.4

# Тестирование
pytest==7.4.3
pytest-django==4.7.0

# Разработка
black==23.12.1
flake8==7.0.0
isort==5.13.2
sqlparse==0.5.1
tzdata==2024.1
django-js-asset==2.2.0

# Tailwind CSS для Django
django-tailwind==3.8.0
django-browser-reload==1.12.1

# Платежные системы и донаты
yookassa==2.3.2
requests==2.31.0
qrcode==7.4.2

# AI-Ассистент и автоматизация
gigachat>=0.1.15
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.10
# django-q2 удалён: заменён Celery
# croniter удалён: больше не нужен (Django-Q удалён)
python-dateutil>=2.8.2

# ML для оптимизации рекламы
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Парсинг видео и аудио (ЭТАП 2)
youtube-transcript-api>=0.6.2
pytube>=15.0.0
SpeechRecognition>=3.10.0
pydub>=0.25.1

# Социальные сети (Sozseti App)
vk-api==11.9.9
httpx>=0.27.0
matplotlib>=3.7.0

# Будущие интеграции (Фаза 3):
# pinterest-api-sdk==0.2.0

# Gunicorn для продакшена
gunicorn>=21.2.0
